uupdated delete course, data seeder, several fixes

+++ server/helpers/seeder.js
@@ -1,12 +1,15 @@
-const readline = require('readline')
-const User = require('../models/User.js')
-const Order = require('../models/Order.js')
-const Course = require('../models/Course.js')
-const CourseProgress = require('../models/CourseProgress.js')
-const StudentCourses = require('../models/StudentCourses.js')
-const users = require('../helpers/data/users.js')
-const courses = require('../helpers/data/courses.js')
+const User = require("../models/User.js")
+const Order = require("../models/Order.js")
+const Course = require("../models/Course.js")
+const CourseProgress = require("../models/CourseProgress.js")
+const StudentCourses = require("../models/StudentCourses.js")
+const courses = require("../helpers/data/courses.js")
+const users = require("../helpers/data/users.js")
+const readline = require("readline")
 
+
+require('dotenv').config()
+require('mongoose').connect(process.env.MONGODB_URI)
 
 const rl = readline.createInterface({
     input: process.stdin,
@@ -14,12 +17,8 @@ const rl = readline.createInterface({
 })
 
 const confirmAction = (message, callback) => {
-    rl.question(`${message} (y/n): `, (answer) => {
+    rl.question(`${message} (y/n):`, (answer) => {
         if (answer.toLowerCase() === 'y') {
-            require('dotenv').config()
-            require('mongoose').connect(process.env.MONGODB_URI)
-                .then(() => console.log('>>> Connected to MongoDB'))
-                .catch((err) => console.log(err))
             callback()
         } else {
             process.exit()
@@ -29,15 +28,32 @@ const confirmAction = (message, callback) => {
 
 const importData = async () => {
     try {
-        await User.deleteMany()
-        await Course.deleteMany()
+        const userCount = await User.countDocuments()
+        const courseCount = await Course.countDocuments()
 
-        await User.insertMany(users)
-        await Course.insertMany(courses)
-        // upload videos & images
+        if (userCount > 0 || courseCount > 0) {
+            confirmAction('❗ Are you sure you want to import new data? This will overwrite existing data', async () => {
+                try {
+                    await User.deleteMany()
+                    await Course.deleteMany()
 
-        console.log('>>> DATA IMPORTED')
-        process.exit()
+                    await User.insertMany(users)
+                    await Course.insertMany(courses)
+
+                    console.log("✔️  DATA IMPORTED")
+                    process.exit()
+                } catch (err) {
+                    console.log(err)
+                    process.exit(1)
+                }
+            })
+        } else {
+            await User.insertMany(users)
+            await Course.insertMany(courses)
+
+            console.log("✔️  DATA IMPORTED")
+            process.exit()
+        }
     } catch (err) {
         console.log(err)
         process.exit(1)
@@ -46,14 +62,26 @@ const importData = async () => {
 
 const destroyData = async () => {
     try {
-        await User.deleteMany()
-        await Order.deleteMany()
-        await Course.deleteMany()
-        await StudentCourses.deleteMany()
-        await CourseProgress.deleteMany()
+        const userCount = await User.countDocuments()
+        const courseCount = await Course.countDocuments()
 
-        console.log('>>> DATA DESTROYED')
-        process.exit()
+        if (userCount === 0 && courseCount === 0) {
+            console.log('Database is empty')
+            process.exit()
+            return
+        }
+        confirmAction('❗ Are you sure you want to delete everything from the database?', async () => {
+            confirmAction('❗ This cannot be undone. Proceed? ', async () => {
+                await User.deleteMany()
+                await Order.deleteMany()
+                await Course.deleteMany()
+                await StudentCourses.deleteMany()
+                await CourseProgress.deleteMany()
+        
+                console.log('✔️  DATA DESTROYED')
+                process.exit()
+            })
+        })
     } catch (err) {
         console.log(err)
         process.exit(1)
@@ -61,10 +89,7 @@ const destroyData = async () => {
 }
 
 if (process.argv[2] === '-d') {
-    confirmAction('Are you sure you want to delete all data? This cannot be undone.', destroyData)
+    destroyData() // npm run destroy-data
 } else {
-    confirmAction('Are you sure you want to import new data?', importData)
+    importData() // npm run import-data
 }
-
-// npm run import-data
-// npm run destroy-data
